src/requirements.txt
gmqtt>=0.6.0
aiocoap>=0.4.5
asyncio-mqtt>=0.13.0
paho-mqtt>=1.6.1
requests>=2.28.0
pyyaml>=6.0
psutil>=5.9.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.4.0
reportlab>=3.6.0
cryptography>=3.4.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
